move actions to constants
