--- src/store/actions.ts
@@ -1,1 +1,0 @@
-export const GET_PROJECTS = 'GET_PROJECTS';
