+++ src/store/constants.ts
@@ -2,5 +2,6 @@ export const ADD_PROJECT = 'ADD_PROJECT';
 export const EDIT_PROJECT = 'EDIT_PROJECT';
 export const DELETE_PROJECT = 'DELETE_PROJECT';
 export const SET_PROJECTS = 'SET_PROJECTS';
+export const GET_PROJECTS = 'GET_PROJECTS';
 
 export const NOTIFY = 'NOTIFY';
